ADD interface_data_controler class switch statment added
