--- docs/clubs/Bundesliga.json
@@ -1,290 +1,0 @@
-{
-  "Augsburg": {
-    "name": " Augsburg",
-    "Matches_Played": "34",
-    "Win": "9",
-    "Draw": "9",
-    "Lost": "16",
-    "GF": "45",
-    "GA": "63",
-    "GDiff": "-18",
-    "points": "36",
-    "XG": "42.0",
-    "XGA": "54.8",
-    "XGDiff": "-12.9",
-    "XGDiff_90": "-0.38",
-    "url": "/en/squads/0cdc4311/Augsburg-Stats"
-  },
-  "Bayern Munich": {
-    "name": " Bayern Munich",
-    "Matches_Played": "34",
-    "Win": "26",
-    "Draw": "4",
-    "Lost": "4",
-    "GF": "100",
-    "GA": "32",
-    "GDiff": "+68",
-    "points": "82",
-    "XG": "82.2",
-    "XGA": "34.3",
-    "XGDiff": "+47.8",
-    "XGDiff_90": "+1.41",
-    "url": "/en/squads/054efa67/Bayern-Munich-Stats"
-  },
-  "Dortmund": {
-    "name": " Dortmund",
-    "Matches_Played": "34",
-    "Win": "21",
-    "Draw": "6",
-    "Lost": "7",
-    "GF": "84",
-    "GA": "41",
-    "GDiff": "+43",
-    "points": "69",
-    "XG": "59.1",
-    "XGA": "39.3",
-    "XGDiff": "+19.8",
-    "XGDiff_90": "+0.58",
-    "url": "/en/squads/add600ae/Dortmund-Stats"
-  },
-  "Düsseldorf": {
-    "name": " Düsseldorf",
-    "Matches_Played": "34",
-    "Win": "6",
-    "Draw": "12",
-    "Lost": "16",
-    "GF": "36",
-    "GA": "67",
-    "GDiff": "-31",
-    "points": "30",
-    "XG": "33.0",
-    "XGA": "56.2",
-    "XGDiff": "-23.2",
-    "XGDiff_90": "-0.68",
-    "url": "/en/squads/b1278397/Dusseldorf-Stats"
-  },
-  "Eint Frankfurt": {
-    "name": " Eint Frankfurt",
-    "Matches_Played": "34",
-    "Win": "13",
-    "Draw": "6",
-    "Lost": "15",
-    "GF": "59",
-    "GA": "60",
-    "GDiff": "-1",
-    "points": "45",
-    "XG": "52.5",
-    "XGA": "49.1",
-    "XGDiff": "+3.4",
-    "XGDiff_90": "+0.10",
-    "url": "/en/squads/f0ac8ee6/Eintracht-Frankfurt-Stats"
-  },
-  "Freiburg": {
-    "name": " Freiburg",
-    "Matches_Played": "34",
-    "Win": "13",
-    "Draw": "9",
-    "Lost": "12",
-    "GF": "48",
-    "GA": "47",
-    "GDiff": "+1",
-    "points": "48",
-    "XG": "41.7",
-    "XGA": "60.6",
-    "XGDiff": "-18.9",
-    "XGDiff_90": "-0.56",
-    "url": "/en/squads/a486e511/Freiburg-Stats"
-  },
-  "Hertha BSC": {
-    "name": " Hertha BSC",
-    "Matches_Played": "34",
-    "Win": "11",
-    "Draw": "8",
-    "Lost": "15",
-    "GF": "48",
-    "GA": "59",
-    "GDiff": "-11",
-    "points": "41",
-    "XG": "40.4",
-    "XGA": "51.3",
-    "XGDiff": "-10.9",
-    "XGDiff_90": "-0.32",
-    "url": "/en/squads/2818f8bc/Hertha-BSC-Stats"
-  },
-  "Hoffenheim": {
-    "name": " Hoffenheim",
-    "Matches_Played": "34",
-    "Win": "15",
-    "Draw": "7",
-    "Lost": "12",
-    "GF": "53",
-    "GA": "53",
-    "GDiff": "0",
-    "points": "52",
-    "XG": "49.7",
-    "XGA": "51.8",
-    "XGDiff": "-2.0",
-    "XGDiff_90": "-0.06",
-    "url": "/en/squads/033ea6b8/Hoffenheim-Stats"
-  },
-  "Köln": {
-    "name": " Köln",
-    "Matches_Played": "34",
-    "Win": "10",
-    "Draw": "6",
-    "Lost": "18",
-    "GF": "51",
-    "GA": "69",
-    "GDiff": "-18",
-    "points": "36",
-    "XG": "47.1",
-    "XGA": "51.5",
-    "XGDiff": "-4.4",
-    "XGDiff_90": "-0.13",
-    "url": "/en/squads/bc357bf7/Koln-Stats"
-  },
-  "Leverkusen": {
-    "name": " Leverkusen",
-    "Matches_Played": "34",
-    "Win": "19",
-    "Draw": "6",
-    "Lost": "9",
-    "GF": "61",
-    "GA": "44",
-    "GDiff": "+17",
-    "points": "63",
-    "XG": "56.6",
-    "XGA": "45.1",
-    "XGDiff": "+11.6",
-    "XGDiff_90": "+0.34",
-    "url": "/en/squads/c7a9f859/Bayer-Leverkusen-Stats"
-  },
-  "M'Gladbach": {
-    "name": " M'Gladbach",
-    "Matches_Played": "34",
-    "Win": "20",
-    "Draw": "5",
-    "Lost": "9",
-    "GF": "66",
-    "GA": "40",
-    "GDiff": "+26",
-    "points": "65",
-    "XG": "66.2",
-    "XGA": "42.4",
-    "XGDiff": "+23.9",
-    "XGDiff_90": "+0.70",
-    "url": "/en/squads/32f3ee20/Monchengladbach-Stats"
-  },
-  "Mainz 05": {
-    "name": " Mainz 05",
-    "Matches_Played": "34",
-    "Win": "11",
-    "Draw": "4",
-    "Lost": "19",
-    "GF": "44",
-    "GA": "65",
-    "GDiff": "-21",
-    "points": "37",
-    "XG": "44.2",
-    "XGA": "57.0",
-    "XGDiff": "-12.8",
-    "XGDiff_90": "-0.38",
-    "url": "/en/squads/a224b06a/Mainz-05-Stats"
-  },
-  "Paderborn 07": {
-    "name": " Paderborn 07",
-    "Matches_Played": "34",
-    "Win": "4",
-    "Draw": "8",
-    "Lost": "22",
-    "GF": "37",
-    "GA": "74",
-    "GDiff": "-37",
-    "points": "20",
-    "XG": "38.5",
-    "XGA": "67.4",
-    "XGDiff": "-28.9",
-    "XGDiff_90": "-0.85",
-    "url": "/en/squads/d9f93f02/Paderborn-07-Stats"
-  },
-  "RB Leipzig": {
-    "name": " RB Leipzig",
-    "Matches_Played": "34",
-    "Win": "18",
-    "Draw": "12",
-    "Lost": "4",
-    "GF": "81",
-    "GA": "37",
-    "GDiff": "+44",
-    "points": "66",
-    "XG": "70.7",
-    "XGA": "36.3",
-    "XGDiff": "+34.4",
-    "XGDiff_90": "+1.01",
-    "url": "/en/squads/acbb6a5b/RB-Leipzig-Stats"
-  },
-  "Schalke 04": {
-    "name": " Schalke 04",
-    "Matches_Played": "34",
-    "Win": "9",
-    "Draw": "12",
-    "Lost": "13",
-    "GF": "38",
-    "GA": "58",
-    "GDiff": "-20",
-    "points": "39",
-    "XG": "35.4",
-    "XGA": "49.8",
-    "XGDiff": "-14.4",
-    "XGDiff_90": "-0.42",
-    "url": "/en/squads/c539e393/Schalke-04-Stats"
-  },
-  "Union Berlin": {
-    "name": " Union Berlin",
-    "Matches_Played": "34",
-    "Win": "12",
-    "Draw": "5",
-    "Lost": "17",
-    "GF": "41",
-    "GA": "58",
-    "GDiff": "-17",
-    "points": "41",
-    "XG": "38.4",
-    "XGA": "51.5",
-    "XGDiff": "-13.1",
-    "XGDiff_90": "-0.39",
-    "url": "/en/squads/7a41008f/Union-Berlin-Stats"
-  },
-  "Werder Bremen": {
-    "name": " Werder Bremen",
-    "Matches_Played": "34",
-    "Win": "8",
-    "Draw": "7",
-    "Lost": "19",
-    "GF": "42",
-    "GA": "69",
-    "GDiff": "-27",
-    "points": "31",
-    "XG": "41.4",
-    "XGA": "48.0",
-    "XGDiff": "-6.6",
-    "XGDiff_90": "-0.20",
-    "url": "/en/squads/62add3bf/Werder-Bremen-Stats"
-  },
-  "Wolfsburg": {
-    "name": " Wolfsburg",
-    "Matches_Played": "34",
-    "Win": "13",
-    "Draw": "10",
-    "Lost": "11",
-    "GF": "48",
-    "GA": "46",
-    "GDiff": "+2",
-    "points": "49",
-    "XG": "47.3",
-    "XGA": "40.0",
-    "XGDiff": "+7.3",
-    "XGDiff_90": "+0.21",
-    "url": "/en/squads/4eaa11d7/Wolfsburg-Stats"
-  }
-}
