--- docs/clubs/La liga.json
@@ -1,1 +1,0 @@
-{"Amiens":{"name":" Amiens","Matches_Played":"28","Win":"4","Draw":"11","Lost":"13","GF":"31","GA":"50","GDiff":"-19","points":"23","XG":"0.82","XGA":"26.9","XGDiff":"35.5","XGDiff_90":"-8.6","url":"/en/squads/25622401/2019-2020/Amiens-Stats"},"Angers":{"name":" Angers","Matches_Played":"28","Win":"11","Draw":"6","Lost":"11","GF":"28","GA":"33","GDiff":"-5","points":"39","XG":"1.39","XGA":"27.0","XGDiff":"28.6","XGDiff_90":"-1.6","url":"/en/squads/69236f98/2019-2020/Angers-Stats"},"Bordeaux":{"name":" Bordeaux","Matches_Played":"28","Win":"9","Draw":"10","Lost":"9","GF":"40","GA":"34","GDiff":"+6","points":"37","XG":"1.32","XGA":"30.1","XGDiff":"33.5","XGDiff_90":"-3.4","url":"/en/squads/123f3efe/2019-2020/Bordeaux-Stats"},"Brest":{"name":" Brest","Matches_Played":"28","Win":"8","Draw":"10","Lost":"10","GF":"34","GA":"37","GDiff":"-3","points":"34","XG":"1.21","XGA":"30.2","XGDiff":"40.5","XGDiff_90":"-10.3","url":"/en/squads/fb08dbb3/2019-2020/Brest-Stats"},"Dijon":{"name":" Dijon","Matches_Played":"28","Win":"7","Draw":"9","Lost":"12","GF":"27","GA":"37","GDiff":"-10","points":"30","XG":"1.07","XGA":"28.7","XGDiff":"43.2","XGDiff_90":"-14.5","url":"/en/squads/8dfb7350/2019-2020/Dijon-Stats"},"Lille":{"name":" Lille","Matches_Played":"28","Win":"15","Draw":"4","Lost":"9","GF":"35","GA":"27","GDiff":"+8","points":"49","XG":"1.75","XGA":"39.6","XGDiff":"29.6","XGDiff_90":"+10.0","url":"/en/squads/cb188c0c/2019-2020/Lille-Stats"},"Lyon":{"name":" Lyon","Matches_Played":"28","Win":"11","Draw":"7","Lost":"10","GF":"42","GA":"27","GDiff":"+15","points":"40","XG":"1.43","XGA":"37.2","XGDiff":"25.9","XGDiff_90":"+11.4","url":"/en/squads/d53c0b06/2019-2020/Lyon-Stats"},"Marseille":{"name":" Marseille","Matches_Played":"28","Win":"16","Draw":"8","Lost":"4","GF":"41","GA":"29","GDiff":"+12","points":"56","XG":"2.00","XGA":"34.0","XGDiff":"30.7","XGDiff_90":"+3.3","url":"/en/squads/5725cc7b/2019-2020/Marseille-Stats"},"Metz":{"name":" Metz","Matches_Played":"28","Win":"8","Draw":"10","Lost":"10","GF":"27","GA":"35","GDiff":"-8","points":"34","XG":"1.21","XGA":"27.6","XGDiff":"34.6","XGDiff_90":"-7.0","url":"/en/squads/f83960ae/2019-2020/Metz-Stats"},"Monaco":{"name":" Monaco","Matches_Played":"28","Win":"11","Draw":"7","Lost":"10","GF":"44","GA":"44","GDiff":"0","points":"40","XG":"1.43","XGA":"48.0","XGDiff":"39.7","XGDiff_90":"+8.3","url":"/en/squads/fd6114db/2019-2020/Monaco-Stats"},"Montpellier":{"name":" Montpellier","Matches_Played":"28","Win":"11","Draw":"7","Lost":"10","GF":"35","GA":"34","GDiff":"+1","points":"40","XG":"1.43","XGA":"31.8","XGDiff":"34.1","XGDiff_90":"-2.2","url":"/en/squads/281b0e73/2019-2020/Montpellier-Stats"},"Nantes":{"name":" Nantes","Matches_Played":"28","Win":"11","Draw":"4","Lost":"13","GF":"28","GA":"31","GDiff":"-3","points":"37","XG":"1.32","XGA":"27.8","XGDiff":"33.4","XGDiff_90":"-5.6","url":"/en/squads/d7a486cd/2019-2020/Nantes-Stats"},"Nice":{"name":" Nice","Matches_Played":"28","Win":"11","Draw":"8","Lost":"9","GF":"41","GA":"38","GDiff":"+3","points":"41","XG":"1.46","XGA":"34.9","XGDiff":"42.3","XGDiff_90":"-7.4","url":"/en/squads/132ebc33/2019-2020/Nice-Stats"},"Nîmes":{"name":" Nîmes","Matches_Played":"28","Win":"7","Draw":"6","Lost":"15","GF":"29","GA":"44","GDiff":"-15","points":"27","XG":"0.96","XGA":"29.7","XGDiff":"43.8","XGDiff_90":"-14.1","url":"/en/squads/1cbf5f9e/2019-2020/Nimes-Stats"},"Paris S-G":{"name":" Paris S-G","Matches_Played":"27","Win":"22","Draw":"2","Lost":"3","GF":"75","GA":"24","GDiff":"+51","points":"68","XG":"2.52","XGA":"71.0","XGDiff":"22.6","XGDiff_90":"+48.4","url":"/en/squads/e2d8892c/2019-2020/Paris-Saint-Germain-Stats"},"Reims":{"name":" Reims","Matches_Played":"28","Win":"10","Draw":"11","Lost":"7","GF":"26","GA":"21","GDiff":"+5","points":"41","XG":"1.46","XGA":"28.2","XGDiff":"26.4","XGDiff_90":"+1.9","url":"/en/squads/7fdd64e0/2019-2020/Reims-Stats"},"Rennes":{"name":" Rennes","Matches_Played":"28","Win":"15","Draw":"5","Lost":"8","GF":"38","GA":"24","GDiff":"+14","points":"50","XG":"1.79","XGA":"36.6","XGDiff":"29.3","XGDiff_90":"+7.3","url":"/en/squads/b3072e00/2019-2020/Rennes-Stats"},"Saint-Étienne":{"name":" Saint-Étienne","Matches_Played":"28","Win":"8","Draw":"6","Lost":"14","GF":"29","GA":"45","GDiff":"-16","points":"30","XG":"1.07","XGA":"29.0","XGDiff":"35.2","XGDiff_90":"-6.3","url":"/en/squads/d298ef2c/2019-2020/Saint-Etienne-Stats"},"Strasbourg":{"name":" Strasbourg","Matches_Played":"27","Win":"11","Draw":"5","Lost":"11","GF":"32","GA":"32","GDiff":"0","points":"38","XG":"1.41","XGA":"34.9","XGDiff":"30.7","XGDiff_90":"+4.1","url":"/en/squads/c0d3eab4/2019-2020/Strasbourg-Stats"},"Toulouse":{"name":" Toulouse","Matches_Played":"28","Win":"3","Draw":"4","Lost":"21","GF":"22","GA":"58","GDiff":"-36","points":"13","XG":"0.46","XGA":"30.4","XGDiff":"44.1","XGDiff_90":"-13.7","url":"/en/squads/3f8c4b5f/2019-2020/Toulouse-Stats"}}
