--- docs/clubs/PL.json
@@ -1,1 +1,0 @@
-{"Arsenal":{"name":" Arsenal","Matches_Played":"38","Win":"14","Draw":"14","Lost":"10","GF":"56","GA":"48","GDiff":"+8","points":"56","XG":"49.1","XGA":"56.1","XGDiff":"-7.1","XGDiff_90":"-0.19","url":"/en/squads/18bb7c10/Arsenal-Stats"},"Aston Villa":{"name":" Aston Villa","Matches_Played":"38","Win":"9","Draw":"8","Lost":"21","GF":"41","GA":"67","GDiff":"-26","points":"35","XG":"40.2","XGA":"65.8","XGDiff":"-25.6","XGDiff_90":"-0.67","url":"/en/squads/8602292d/Aston-Villa-Stats"},"Bournemouth":{"name":" Bournemouth","Matches_Played":"38","Win":"9","Draw":"7","Lost":"22","GF":"40","GA":"65","GDiff":"-25","points":"34","XG":"42.5","XGA":"56.7","XGDiff":"-14.2","XGDiff_90":"-0.37","url":"/en/squads/4ba7cbea/Bournemouth-Stats"},"Brighton":{"name":" Brighton","Matches_Played":"38","Win":"9","Draw":"14","Lost":"15","GF":"39","GA":"54","GDiff":"-15","points":"41","XG":"40.8","XGA":"55.1","XGDiff":"-14.3","XGDiff_90":"-0.38","url":"/en/squads/d07537b9/Brighton-and-Hove-Albion-Stats"},"Burnley":{"name":" Burnley","Matches_Played":"38","Win":"15","Draw":"9","Lost":"14","GF":"43","GA":"50","GDiff":"-7","points":"54","XG":"43.9","XGA":"47.8","XGDiff":"-3.9","XGDiff_90":"-0.10","url":"/en/squads/943e8050/Burnley-Stats"},"Chelsea":{"name":" Chelsea","Matches_Played":"38","Win":"20","Draw":"6","Lost":"12","GF":"69","GA":"54","GDiff":"+15","points":"66","XG":"66.6","XGA":"37.1","XGDiff":"+29.6","XGDiff_90":"+0.78","url":"/en/squads/cff3d9bb/Chelsea-Stats"},"Crystal Palace":{"name":" Crystal Palace","Matches_Played":"38","Win":"11","Draw":"10","Lost":"17","GF":"31","GA":"50","GDiff":"-19","points":"43","XG":"33.7","XGA":"50.9","XGDiff":"-17.1","XGDiff_90":"-0.45","url":"/en/squads/47c64c55/Crystal-Palace-Stats"},"Everton":{"name":" Everton","Matches_Played":"38","Win":"13","Draw":"10","Lost":"15","GF":"44","GA":"56","GDiff":"-12","points":"49","XG":"48.3","XGA":"47.3","XGDiff":"+1.0","XGDiff_90":"+0.03","url":"/en/squads/d3fd31cc/Everton-Stats"},"Leicester City":{"name":" Leicester City","Matches_Played":"38","Win":"18","Draw":"8","Lost":"12","GF":"67","GA":"41","GDiff":"+26","points":"62","XG":"60.8","XGA":"44.8","XGDiff":"+16.0","XGDiff_90":"+0.42","url":"/en/squads/a2d435b3/Leicester-City-Stats"},"Liverpool":{"name":" Liverpool","Matches_Played":"38","Win":"32","Draw":"3","Lost":"3","GF":"85","GA":"33","GDiff":"+52","points":"99","XG":"70.4","XGA":"40.0","XGDiff":"+30.4","XGDiff_90":"+0.80","url":"/en/squads/822bd0ba/Liverpool-Stats"},"Manchester City":{"name":" Manchester City","Matches_Played":"38","Win":"26","Draw":"3","Lost":"9","GF":"102","GA":"35","GDiff":"+67","points":"81","XG":"93.1","XGA":"34.3","XGDiff":"+58.8","XGDiff_90":"+1.55","url":"/en/squads/b8fd03ef/Manchester-City-Stats"},"Manchester Utd":{"name":" Manchester Utd","Matches_Played":"38","Win":"18","Draw":"12","Lost":"8","GF":"66","GA":"36","GDiff":"+30","points":"66","XG":"59.8","XGA":"37.3","XGDiff":"+22.6","XGDiff_90":"+0.59","url":"/en/squads/19538871/Manchester-United-Stats"},"Newcastle Utd":{"name":" Newcastle Utd","Matches_Played":"38","Win":"11","Draw":"11","Lost":"16","GF":"38","GA":"58","GDiff":"-20","points":"44","XG":"33.0","XGA":"58.5","XGDiff":"-25.5","XGDiff_90":"-0.67","url":"/en/squads/b2b47a98/Newcastle-United-Stats"},"Norwich City":{"name":" Norwich City","Matches_Played":"38","Win":"5","Draw":"6","Lost":"27","GF":"26","GA":"75","GDiff":"-49","points":"21","XG":"38.2","XGA":"61.9","XGDiff":"-23.7","XGDiff_90":"-0.62","url":"/en/squads/1c781004/Norwich-City-Stats"},"Sheffield Utd":{"name":" Sheffield Utd","Matches_Played":"38","Win":"14","Draw":"12","Lost":"12","GF":"39","GA":"39","GDiff":"0","points":"54","XG":"41.5","XGA":"47.9","XGDiff":"-6.4","XGDiff_90":"-0.17","url":"/en/squads/1df6b87e/Sheffield-United-Stats"},"Southampton":{"name":" Southampton","Matches_Played":"38","Win":"15","Draw":"7","Lost":"16","GF":"51","GA":"60","GDiff":"-9","points":"52","XG":"53.9","XGA":"53.2","XGDiff":"+0.7","XGDiff_90":"+0.02","url":"/en/squads/33c895d4/Southampton-Stats"},"Tottenham":{"name":" Tottenham","Matches_Played":"38","Win":"16","Draw":"11","Lost":"11","GF":"61","GA":"47","GDiff":"+14","points":"59","XG":"45.5","XGA":"51.9","XGDiff":"-6.3","XGDiff_90":"-0.17","url":"/en/squads/361ca564/Tottenham-Hotspur-Stats"},"Watford":{"name":" Watford","Matches_Played":"38","Win":"8","Draw":"10","Lost":"20","GF":"36","GA":"64","GDiff":"-28","points":"34","XG":"45.6","XGA":"57.2","XGDiff":"-11.7","XGDiff_90":"-0.31","url":"/en/squads/2abfe087/Watford-Stats"},"West Ham":{"name":" West Ham","Matches_Played":"38","Win":"10","Draw":"9","Lost":"19","GF":"49","GA":"62","GDiff":"-13","points":"39","XG":"46.1","XGA":"61.9","XGDiff":"-15.8","XGDiff_90":"-0.42","url":"/en/squads/7c21e445/West-Ham-United-Stats"},"Wolves":{"name":" Wolves","Matches_Played":"38","Win":"15","Draw":"14","Lost":"9","GF":"51","GA":"40","GDiff":"+11","points":"59","XG":"47.0","XGA":"34.4","XGDiff":"+12.5","XGDiff_90":"+0.33","url":"/en/squads/8cec06e1/Wolverhampton-Wanderers-Stats"}}
