--- docs/clubs/Seari A.json
@@ -1,1 +1,0 @@
-{"Atalanta":{"name":" Atalanta","Matches_Played":"38","Win":"23","Draw":"9","Lost":"6","GF":"98","GA":"48","GDiff":"+50","points":"78","XG":"82.1","XGA":"41.3","XGDiff":"+40.8","XGDiff_90":"+1.07","url":"/en/squads/922493f3/Atalanta-Stats"},"Bologna":{"name":" Bologna","Matches_Played":"38","Win":"12","Draw":"11","Lost":"15","GF":"52","GA":"65","GDiff":"-13","points":"47","XG":"53.2","XGA":"54.2","XGDiff":"-0.9","XGDiff_90":"-0.02","url":"/en/squads/1d8099f8/Bologna-Stats"},"Brescia":{"name":" Brescia","Matches_Played":"38","Win":"6","Draw":"7","Lost":"25","GF":"35","GA":"79","GDiff":"-44","points":"25","XG":"42.6","XGA":"77.6","XGDiff":"-35.0","XGDiff_90":"-0.92","url":"/en/squads/4ef57aeb/Brescia-Stats"},"Cagliari":{"name":" Cagliari","Matches_Played":"38","Win":"11","Draw":"12","Lost":"15","GF":"52","GA":"56","GDiff":"-4","points":"45","XG":"41.2","XGA":"57.3","XGDiff":"-16.1","XGDiff_90":"-0.42","url":"/en/squads/c4260e09/Cagliari-Stats"},"Fiorentina":{"name":" Fiorentina","Matches_Played":"38","Win":"12","Draw":"13","Lost":"13","GF":"51","GA":"48","GDiff":"+3","points":"49","XG":"54.5","XGA":"48.9","XGDiff":"+5.6","XGDiff_90":"+0.15","url":"/en/squads/421387cf/Fiorentina-Stats"},"Genoa":{"name":" Genoa","Matches_Played":"38","Win":"10","Draw":"9","Lost":"19","GF":"47","GA":"73","GDiff":"-26","points":"39","XG":"47.5","XGA":"59.5","XGDiff":"-12.0","XGDiff_90":"-0.32","url":"/en/squads/658bf2de/Genoa-Stats"},"Hellas Verona":{"name":" Hellas Verona","Matches_Played":"38","Win":"12","Draw":"13","Lost":"13","GF":"47","GA":"51","GDiff":"-4","points":"49","XG":"48.9","XGA":"52.4","XGDiff":"-3.4","XGDiff_90":"-0.09","url":"/en/squads/0e72edf2/Hellas-Verona-Stats"},"Inter":{"name":" Inter","Matches_Played":"38","Win":"24","Draw":"10","Lost":"4","GF":"81","GA":"36","GDiff":"+45","points":"82","XG":"72.1","XGA":"40.6","XGDiff":"+31.5","XGDiff_90":"+0.83","url":"/en/squads/d609edc0/Internazionale-Stats"},"Juventus":{"name":" Juventus","Matches_Played":"38","Win":"26","Draw":"5","Lost":"7","GF":"76","GA":"43","GDiff":"+33","points":"83","XG":"68.9","XGA":"41.9","XGDiff":"+27.0","XGDiff_90":"+0.71","url":"/en/squads/e0652b02/Juventus-Stats"},"Lazio":{"name":" Lazio","Matches_Played":"38","Win":"24","Draw":"6","Lost":"8","GF":"79","GA":"42","GDiff":"+37","points":"78","XG":"68.0","XGA":"49.5","XGDiff":"+18.6","XGDiff_90":"+0.49","url":"/en/squads/7213da33/Lazio-Stats"},"Lecce":{"name":" Lecce","Matches_Played":"38","Win":"9","Draw":"8","Lost":"21","GF":"52","GA":"85","GDiff":"-33","points":"35","XG":"53.8","XGA":"84.1","XGDiff":"-30.3","XGDiff_90":"-0.80","url":"/en/squads/ffcbe334/Lecce-Stats"},"Milan":{"name":" Milan","Matches_Played":"38","Win":"19","Draw":"9","Lost":"10","GF":"63","GA":"46","GDiff":"+17","points":"66","XG":"60.1","XGA":"47.7","XGDiff":"+12.5","XGDiff_90":"+0.33","url":"/en/squads/dc56fe14/Milan-Stats"},"Napoli":{"name":" Napoli","Matches_Played":"38","Win":"18","Draw":"8","Lost":"12","GF":"61","GA":"50","GDiff":"+11","points":"62","XG":"62.5","XGA":"39.4","XGDiff":"+23.1","XGDiff_90":"+0.61","url":"/en/squads/d48ad4ff/Napoli-Stats"},"Parma":{"name":" Parma","Matches_Played":"38","Win":"14","Draw":"7","Lost":"17","GF":"56","GA":"57","GDiff":"-1","points":"49","XG":"46.6","XGA":"62.1","XGDiff":"-15.5","XGDiff_90":"-0.41","url":"/en/squads/eab4234c/Parma-Stats"},"Roma":{"name":" Roma","Matches_Played":"38","Win":"21","Draw":"7","Lost":"10","GF":"77","GA":"51","GDiff":"+26","points":"70","XG":"72.6","XGA":"48.3","XGDiff":"+24.3","XGDiff_90":"+0.64","url":"/en/squads/cf74a709/Roma-Stats"},"SPAL":{"name":" SPAL","Matches_Played":"38","Win":"5","Draw":"5","Lost":"28","GF":"27","GA":"77","GDiff":"-50","points":"20","XG":"39.0","XGA":"67.3","XGDiff":"-28.2","XGDiff_90":"-0.74","url":"/en/squads/1d2fe027/SPAL-Stats"},"Sampdoria":{"name":" Sampdoria","Matches_Played":"38","Win":"12","Draw":"6","Lost":"20","GF":"48","GA":"65","GDiff":"-17","points":"42","XG":"51.2","XGA":"53.4","XGDiff":"-2.1","XGDiff_90":"-0.06","url":"/en/squads/8ff9e3b3/Sampdoria-Stats"},"Sassuolo":{"name":" Sassuolo","Matches_Played":"38","Win":"14","Draw":"9","Lost":"15","GF":"69","GA":"63","GDiff":"+6","points":"51","XG":"55.9","XGA":"64.3","XGDiff":"-8.3","XGDiff_90":"-0.22","url":"/en/squads/e2befd26/Sassuolo-Stats"},"Torino":{"name":" Torino","Matches_Played":"38","Win":"11","Draw":"7","Lost":"20","GF":"46","GA":"68","GDiff":"-22","points":"40","XG":"42.3","XGA":"62.6","XGDiff":"-20.2","XGDiff_90":"-0.53","url":"/en/squads/105360fe/Torino-Stats"},"Udinese":{"name":" Udinese","Matches_Played":"38","Win":"12","Draw":"9","Lost":"17","GF":"37","GA":"51","GDiff":"-14","points":"45","XG":"40.7","XGA":"51.9","XGDiff":"-11.2","XGDiff_90":"-0.30","url":"/en/squads/04eea015/Udinese-Stats"}}
