+++ lib/constants.rb
@@ -32,6 +32,7 @@ PLAYER_ATT_ATTRIB = %i[
   Assist_per_90mins
   G_plus_A_per_90mins
 ].freeze
+
 PLAYER_DEF_ATTRIB = %i[
   Tackles
   Tackles_won
@@ -41,6 +42,7 @@ PLAYER_DEF_ATTRIB = %i[
   Clearances
   Errors
 ].freeze
+
 PLAYER_DEF_SELECTED_INDEX = [4, 5, 19, 23, 24, 25, 26].freeze
 PLAYER_GK_ATTRIB = %i[
   Goals_conceded
@@ -49,6 +51,15 @@ PLAYER_GK_ATTRIB = %i[
   Save_%
   Clean_sheet
 ].freeze
+
+CLUB_WEBSITE = {
+  English_Premier_League: 'https://fbref.com/en/comps/9/Premier-League-Stats',
+  Spain_La_Liga: 'https://fbref.com/en/comps/12/La-Liga-Stats',
+  Germany_Bundesliga: 'https://fbref.com/en/comps/20/Bundesliga-Stats',
+  Italian_Serie_A: 'https://fbref.com/en/comps/11/Serie-A-Stats',
+  France_Ligue1: 'https://fbref.com/en/comps/13/3243/2019-2020-Ligue-1-Stats'
+}.freeze
+
 PLAYER_GK_SELECTED_INDEX = [6, 7, 9, 10, 14].freeze
 TEAM_DIR = './docs/clubs/'.freeze
 PLAYERS_DIR = './docs/players/'.freeze
